// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Articulo.ts
export interface Articulo {
  claveArtSisVB?: string;
  claveArticulo: string;
  generico: boolean;
  dosPiezas: boolean;
  revision: string;
  verVentas: string;
  idCliente: string;
  razonSocial: string;
  tipoProducto: string;
  descripcion: string;
  noPartes: string;
  revClientes: string;
  pulgadas: string;
  modelo: string;
  subCodigo: number;
  resistencia: string;
  tipoCaja: number;
  tipoUnion: string;
  desarrollo: string;
  flauta: string;
  suaje50: string;
  suaje66: string;
  grabado: number;
  tipoGrabado: string;
  presentacion: string;
  autMaestraRes: boolean;
  hotMelt: boolean;
  embHotMelt: boolean;
  matPaletizado: boolean;
  esPrioritario: boolean;
  paletizadoEspecial: boolean;
  stockMax: number;
  stockMin: number;
  clavePreparacion: string;
  maestraNo: string;
  colorEtiqueta: number;
  tipoPallet: string;
  palletCargados: number;
  totalPzasCargadas: number;
  tipoPaletizado: number;
  llenadoPallet: number;
  largoInt: number;
  anchoInt: number;
  altoInt: number;
  largoDes: number;
  anchoDes: number;
  dcDes: string;
  areaUnitaria: number;
  pesoUnitario: number;
  porsDesperdicio: number;
  bctRequeridoKgf: number;
  bctRequeridoLbf: number;
  bctObtenidoKgf: number;
  bctObtenidoLbf: number;
  ectEsperado: number;
  color1: number;
  color2: number;
  color3: number;
  color4: number;
  piezasXBulto: number;
  bultosXCama: number;
  camasXPallet: number;
  piezasXPallet: number;
  inventario: string;
  entrada: number;
  tratamientoTermico: boolean;
  bolsaNegra: boolean;
  supModDatosPalArt: boolean;
  anio: number;
  proceso: number;
  proveedor: string;
  nombreProveedor: string;
  escalaCaja: number;
  destino: string;
  tipoIdentificacion: string;
  valorCodigo: string;
  acomodoSPallet: number;
  acomodoUnidad: number;
  tipoPapel: Array<TipoPapel>;
  etiqueta: Array<Etiqueta>;
}
export class Articulo {
  claveArticulo: string;
  generico: boolean;
  dosPiezas: boolean;
  revision: string;
  verVentas: string;
  idCliente: string;
  razonSocial: string;
  tipoProducto: string;
  descripcion: string;
  noPartes: string;
  revClientes: string;
  pulgadas: string;
  modelo: string;
  subCodigo: number;
  resistencia: string;
  tipoCaja: number;
  tipoUnion: string;
  desarrollo: string;
  flauta: string;
  suaje50: string;
  suaje66: string;
  grabado: number;
  tipoGrabado: string;
  presentacion: string;
  autMaestraRes: boolean;
  hotMelt: boolean;
  embHotMelt: boolean;
  matPaletizado: boolean;
  esPrioritario: boolean;
  paletizadoEspecial: boolean;
  stockMax: number;
  stockMin: number;
  clavePreparacion: string;
  maestraNo: string;
  colorEtiqueta: number;
  tipoPallet: string;
  palletCargados: number;
  totalPzasCargadas: number;
  tipoPaletizado: number;
  llenadoPallet: number;
  largoInt: number;
  anchoInt: number;
  altoInt: number;
  largoDes: number;
  anchoDes: number;
  dcDes: string;
  areaUnitaria: number;
  pesoUnitario: number;
  porsDesperdicio: number;
  bctRequeridoKgf: number;
  bctRequeridoLbf: number;
  bctObtenidoKgf: number;
  bctObtenidoLbf: number;
  ectEsperado: number;
  color1: number;
  color2: number;
  color3: number;
  color4: number;
  piezasXBulto: number;
  bultosXCama: number;
  camasXPallet: number;
  piezasXPallet: number;
  inventario: string;
  entrada: number;
  tratamientoTermico: boolean;
  bolsaNegra: boolean;
  supModDatosPalArt: boolean;
  anio: number;
  proceso: number;
  proveedor: string;
  nombreProveedor: string;
  escalaCaja: number;
  destino: string;
  tipoIdentificacion: string;
  valorCodigo: string;
  acomodoSPallet: number;
  acomodoUnidad: number;
  tipoPapel: Array<TipoPapel>;
  etiqueta: Array<Etiqueta>;
  ListaArticulos: ListarArticulos [] = [];
}
export interface FiltrosArt {
  filtro: string;
  TipoProducto: string;
  Resistencia: string;
  Suaje: string;
  idCliente: string;
  generico: boolean;
  sinBackFrom: boolean;
  Suspendido: boolean;
  Estatus: boolean;
  ComponentePrin?: string;
}
export interface ListarArticulos {
  claveArtSisVB?: string;
  ClaveArticulo?: string;
  Descripcion?: string;
  SubCodigo?: string;
  Diseno?: string;
  DisenoA?: string;
  LargoInterior?: number;
  AnchoInterior?: number;
  AltoInterior?: number;
  Resistencia?: string;
  Flauta?: string;
  LargoDesarrollo?: number;
  AnchoDesarrollo?: number;
  ClavePreparacion?: string;
  Color1?: string;
  Color2?: string;
  Color3?: string;
  Color4?: string;
  PorDesperdicio?: number;
  Suaje?: string;
  Dado?: string;
  Papel?: string;
  HotMelt?: boolean;
  AreaUnitaria?: number;
  PesoUnitario?: number;
  Fecha?: Date;
  PiezasXCorte?: string;
  ConScore?: string;
  PiezasXHoja?: number;
  Parafinado?: number;
  PiezasXbulto?: number;
  BultosXcama?: number;
  CamasXPallet?: number;
  troquel?: boolean;
  status?: boolean;
  FechaBaja?: Date;
  largohoja?: number;
  anchohoja?: number;
  Revision?: number;
  MaquinaEstablecida?: string;
  Suaje2?: string;
  CompresionEsperada?: number;
  Dado50?: string;
  Suaje50?: string;
  PesoEspEnCalidad?: number;
  ImprimeCertificado?: boolean;
  PesoMinCalidad?: number;
  PesoMaxCalidad?: number;
  TipoUnion?: string;
  Bodega?: string;
  Impermeabilizado?: string;
  StringKing?: boolean;
  GramosHotMelt?: number;
  CompresionEsperadaLBF?: number;
  Proceso?: string;
  Pallet?: boolean;
  Cproceso?: string;
  Fecha_Act?: Date;
  Usuario_Act?: string;
  VersionDigital?: number;
  FechaDigital?: Date;
  UsuarioDigital?: string;
  Inventario?: string;
  ECT_Esp?: number;
  NScores?: number;
  SuajePL?: string;
  Grabado?: string;
  TipoGrabado?: string;
  Suaje66?: string;
  CvePresupuestal?: string;
  EmbarcaHotMelt?: boolean;
  Maximo?: number;
  Minimo?: number;
  CostoAdicional?: number;
  Comentarios?: string;
  ClaveProceso?: string;
  Desarrollo?: string;
  HandHold?: boolean;
  ComCertificado?: boolean;
  LiberadoSuaje?: boolean;
  LiberadoGrabado?: boolean;
  FechaMod?: Date;
  Depurado?: boolean;
  PzasxLargo?: number;
  PzasxAncho?: number;
  AumLargo?: number;
  AumAncho?: number;
  SuajeDepurado?: boolean;
  ImpresionDepurada?: boolean;
  Eproceso?: string;
  EProcesoCorr?: string;
  SoloLamina?: number;
  Comentario?: string;
  TipoPallet?: string;
  AplicaCaja?: boolean;
  CodigoCaja?: string;
  AplicaLitho?: boolean;
  CodigoLitho?: string;
  RevisionSolicitud?: string;
  SupModPallet?: boolean;
  AplicaSelloFecheroImp?: boolean;
  IdTipoIdentCBCaja?: string;
  IdTipoIdentCBLitho?: string;
  AvisoCambioCombinacion?: string;
  BCTObtenidoLbf?: string;
  BCTObtenido?: string;
  CotLiner1?: string;
  CotCorrugado1?: string;
  CotLiner2?: string;
  CotCorrugado2?: string;
  CotLiner3?: string;
  CotCorrugado3?: string;
  CotLiner4?: string;
  Usuario_Suspendio?: string;
  Fecha_Suspendio?: string;
  Suspendido?: string;
  PzasCargadas?: string;
  AcomodoSPallet?: string;
  AcomodoUnidad?: string;
  TipoPaletizado?: string;
  llenadoPallet?: string;
  PaletEspecial?: string;
  Acomodo?: string;
  PalletsCargados?: string;
  ColorEtiqueta?: string;
  CompresionEspOtrosLBF?: string;
  EsArticuloPrioritario?: string;
  DirCorrugado?: string;
  CodigoDatoPallet?: string;
  ClaveGlobal?: string;
  CodigoRelCliente?: string;
  Articulos2Piezas?: string;
  PesoM2PapelYGoma?: string;
  PesoUnitProdArt?: string;
  RevCliente?: string;
  Cantidad?: number;
  CantidadOP?: number;
  Linia?: number;
  PxJ?: number;
  TipoCajaDes?: string;
}