// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/AsginacionMaquina.ts
export class AsginacionMaquina {
  maquinaOrden: number;
  claveArticulo: string;
  claveMaquina: string;
  activo: number;
  estatus: number;
  usuarioInsert: string;
  fechaInsert: Date;
  usuarioUpdate: string;
  fechaUpdate: Date;
  usuarioDelete: string;
  fechaDelete: Date;

}